+++ packages/buildtool/index.mjs
@@ -30,7 +30,7 @@ export async function build({
   process.chdir(dir);
   fs.rmSync("dist", { recursive: true, force: true });
 
-  const esbuildPlugin = external => ({
+  const esbuildPlugin = (external, alias = {}) => ({
     name: "esbuild",
     async load(id) {
       const { outputFiles } = await esbuild.build({
@@ -51,6 +51,7 @@ export async function build({
           __NAME__: '"@netless/fastboard"',
           __VERSION__: JSON.stringify(version),
         },
+        alias,
         external: Object.keys({
           ...dependencies,
           ...peerDependencies,
@@ -66,30 +67,61 @@ export async function build({
     },
   });
 
+  // Build dist/index.js
   let start = Date.now();
-  let bundle = await rollup.rollup({
-    input: main,
-    plugins: [esbuildPlugin()],
-    external: [/^[@a-z]/],
-  });
+  {
+    let bundle = await rollup.rollup({
+      input: main,
+      plugins: [esbuildPlugin()],
+      external: [/^[@a-z]/],
+    });
 
-  const esm = bundle.write({
-    file: "dist/index.mjs",
-    format: "es",
-    sourcemap: true,
-  });
+    const esm = bundle.write({
+      file: "dist/index.mjs",
+      format: "es",
+      sourcemap: true,
+    });
 
-  const cjs = bundle.write({
-    file: "dist/index.js",
-    format: "cjs",
-    sourcemap: true,
-    interop: "auto",
-  });
+    const cjs = bundle.write({
+      file: "dist/index.js",
+      format: "cjs",
+      sourcemap: true,
+      interop: "auto",
+    });
 
-  await esm;
-  await cjs;
-  await bundle.close();
-  console.log("Built dist/index.{js|mjs} in", Date.now() - start + "ms");
+    await esm;
+    await cjs;
+    await bundle.close();
+    console.log("Built dist/index.{js|mjs} in", Date.now() - start + "ms");
+  }
+
+  // Build dist/lite.js
+  start = Date.now();
+  if (!name.endsWith("-ui")) {
+    let bundle = await rollup.rollup({
+      input: name.endsWith("-core") ? "src/lite.ts" : main,
+      plugins: [esbuildPlugin([], { "@netless/fastboard-core": "@netless/fastboard-core/lite" })],
+      external: [/^[@a-z]/],
+    });
+
+    const esm = bundle.write({
+      file: "dist/lite.mjs",
+      format: "es",
+      sourcemap: true,
+    });
+
+    const cjs = bundle.write({
+      file: "dist/lite.js",
+      format: "cjs",
+      sourcemap: true,
+      interop: "auto",
+    });
+
+    await esm;
+    await cjs;
+    await bundle.close();
+    console.log("Built dist/lite.{js|mjs} in", Date.now() - start + "ms");
+  }
 
   // Generate dist/index.svelte.mjs for the UI package,
   // this also helps to avoid depending on svelte in fastboard/fastboard-react test.
@@ -112,4 +144,16 @@ export async function build({
   start = Date.now();
   await dts.build(main, "dist/index.d.ts", { exclude: ["svelte", "svelte/internal"] });
   console.log("Built dist/index.d.ts in", Date.now() - start + "ms");
+
+  // Generate dist/lite.d.ts
+  start = Date.now();
+  if (name.endsWith("-core")) {
+    await dts.build("src/lite.ts", "dist/lite.d.ts", { exclude: ["svelte", "svelte/internal"] });
+    console.log("Built dist/lite.d.ts in", Date.now() - start + "ms");
+  } else if (!name.endsWith("-ui")) {
+    let code = fs.readFileSync("dist/index.d.ts", "utf-8");
+    code = code.replace(/@netless\/fastboard-core/g, "@netless/fastboard-core/lite");
+    fs.writeFileSync("dist/lite.d.ts", code);
+    console.log("Built dist/lite.d.ts in", Date.now() - start + "ms");
+  }
 }